put doc-cache behind condition for local dev

+++ docs/generate_docs.sh
@@ -5,9 +5,10 @@ OUTPUT_DIR='../dowhy-docs'
 #
 # Cache existing docs
 #
-[ -d ${OUTPUT_DIR} ] && rm -rf ${OUTPUT_DIR}
-git clone --branch gh-pages https://github.com/py-why/dowhy.git ${OUTPUT_DIR}
-rm -rf ${OUTPUT_DIR}/.git
+if [ ! -f "${OUTPUT_DIR}/index.html" ]; then
+    git clone --branch gh-pages https://github.com/py-why/dowhy.git ${OUTPUT_DIR}
+    rm -rf ${OUTPUT_DIR}/.git
+fi
 
 #
 # Build <0.9 Versions using RTD Theme
